refactor: changes in dtos path
