+++ employee/src/main/java/com/jobby/employee/infraestructure/dtos/objects/created/AddressCreated.java
@@ -1,4 +1,4 @@
-package com.jobby.employee.infraestructure.dto;
+package com.jobby.employee.infraestructure.dtos.objects.created;
 
 import lombok.Getter;
 import lombok.Setter;
